backend/requirements.txt
pyyaml
requests
Flask
flask-cors
